services/generation_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pyyaml
httpx
PyJWT
passlib
jinja2
